Use base class for pagination logic
